requirements.txt
requests==2.18.3
numpy==1.12.0
Pillow==4.2.1
beautifulsoup4==4.6.0